feat: use enum directly for display
